// apps/consumer-experience/src/services/display-rules/types.ts
export enum ComponentName {
  OVERVIEW_PROFILE = 'OVERVIEW_PROFILE',
  OVERVIEW_PREMIUM_LINK = 'OVERVIEW_PREMIUM_LINK',
  OVERVIEW_PREMIUM_DETAILED_VIEW = 'OVERVIEW_PREMIUM_DETAILED_VIEW',
  OVERVIEW_PAYMENT_HISTORY = 'OVERVIEW_PAYMENT_HISTORY',
  OVERVIEW_COVERAGE = 'OVERVIEW_COVERAGE',
  OVERVIEW_ACCOUNT_VALUE = 'OVERVIEW_ACCOUNT_VALUE',
  OVERVIEW_BENEFICIARIES = 'OVERVIEW_BENEFICIARIES',
  OVERVIEW_RIDERS = 'OVERVIEW_RIDERS',
  OVERVIEW_DOCUMENTS = 'OVERVIEW_DOCUMENTS',
  PROFILE_PAYOR_PARTY_ROLES = 'PROFILE_PAYOR_PARTY_ROLES',
  PREMIUM_PAYOR_BACK_URL = 'PREMIUM_PAYOR_BACK_URL',
}